pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memebot"
version = "0.0.1"
description = "A Maubot plugin that promotes images to an external servers when commanded in a Matrix room"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "lucaszischka", email = "lucas_zischka@outlook.com" },
]
keywords = ["maubot", "matrix", "bot", "image", "promotion", "meme"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Chat",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "Pillow>=10.0.0",
    "cryptography>=41.0.0",
    "maubot>=0.5.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "aiohttp>=3.9.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "requests>=2.32.0",
    "types-requests>=2.32.0",
]

[project.urls]
Homepage = "https://github.com/lucaszischka/MemeBot"
Repository = "https://github.com/lucaszischka/MemeBot"
Issues = "https://github.com/lucaszischka/MemeBot/issues"
Documentation = "https://github.com/lucaszischka/MemeBot/blob/main/README.md"

[tool.hatch.build.targets.wheel]
packages = ["MemeBot"]

[tool.hatch.version]
path = "MemeBot/__init__.py"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | archive
  | builds
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "maubot.*",
    "mautrix.*",
    "PIL.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"
